// src/features/products/types.ts
export type Product = {
  id: string;
  name: string;
  price: number; // in IDR
  image: string;
  rating: number; // 0..5
  skinType: string;
  keyIngredients: string[];
  ingredients: string[];
  benefits: string[];
  size: string; // e.g. "100 ml"
  category: string;
  category_id: number;
  product_type_id: number;
  how_to_use: string;
  tokopedia_url?: string | null;
  shopee_url?: string | null;
  review_count?: number;
  click_count?: number;
};

export type Review = {
  id: string;
  user_id: string;
  product_id: string; // UUID string
  rating: number;
  comment: string | null;      // Komentar bisa kosong (null)
  created_at: string;          // Supabase return string, bukan Date

  // Penting! Ini untuk menampung data join dari tabel users
  user?: {
    full_name?: string | null;
    avatar_url?: string | null;
    email?: string | null;
  } | null;
};

export type Cart = {
  id: string;
  user_id: string | null;
  status: 'active' | 'completed' | 'abandoned';
  created_at: string;
};

export type CartItem = {
  id: string;
  cart_id: string;
  product_id: string;
  quantity: number;
  created_at?: string;
  product?: Product;
};